Move Modrinth update to addons
